Move mock-xhr helper back into test directory to fix TypeScript build
The previous commit moved mock-xhr.ts from test/ to test-utils/,
which exposed it to the TypeScript build. Since test-utils/ is not
excluded by tsconfig, the jest.fn() calls caused TS2304 errors
during CI builds. Moving it back into test/ (which is excluded)
fixes the build while keeping the shared helper accessible to both
test files.
